Conexão com  Banco de Dados

+++ projeto/monitoramento/models.py
@@ -1,11 +1,11 @@
 from django.db import models
 
 class Leituras(models.Model):
-    temperatura = models.IntegerField()
+    temperatura = models.FloatField()
     ph = models.IntegerField()
     tds = models.IntegerField()
     data_hora = models.DateTimeField(auto_now_add=True)
 
-def __str__(self):
-    return f"{self.temperatura} C | pH: {self.ph} | TDS: {self.tds} | Horario: {self.data_hora}"
+    def __str__(self):
+        return f"{self.temperatura} C | pH: {self.ph} | TDS: {self.tds} | Horario: {self.data_hora}"
